Refactored server folder structure. Yes, again
