+++ server/internal/constant.go
@@ -1,0 +1,12 @@
+package internal
+
+const minThreshold = -4000000000000000000
+const maxThreshold = +4000000000000000000
+
+var (
+	equal   = "="
+	less    = "<"
+	greater = ">"
+)
+
+var encryptedMessage = "something interesting"
