+++ server/internal/random.go
@@ -1,0 +1,9 @@
+package internal
+
+import "math/rand"
+
+var randomNumber int64
+
+func shuffleRandomNumber() {
+	randomNumber = rand.Int63n(maxThreshold-minThreshold+1) + minThreshold
+}
